#85 ファイルポーリングWorkerを廃止
- app/worker.go を削除
- app/main.go からWorker初期化・起動コードを削除
- buildDiaryPrompt等の共有関数を app/prompt.go に移動
